+++ src/API/Type/BaseFolderIdType.php
@@ -2,6 +2,7 @@
 
 namespace garethp\ews\API\Type;
 
+use garethp\ews\API\Exception;
 use garethp\ews\API\Type;
 
 /**
@@ -12,6 +13,8 @@ use garethp\ews\API\Type;
  */
 class BaseFolderIdType extends Type
 {
-
-
+    public function toArray($getOuterArray = false)
+    {
+        throw new Exception('Cannot use BaseFolderIdType directly');
+    }
 }